remove test code from landing page
