--- src/store/type.ts
@@ -1,21 +1,0 @@
-
-// All store types
-export interface WaitlistEntryData {
-  name: string;
-  email: string;
-  role: string;
-  newsletter: boolean;
-}
-
-export interface WaitlistResponse {
-  waitlist: WaitlistEntryData;
-  createWaitlist: ( { email, name, newsletter, role }: WaitlistEntryData ) => Promise<boolean>;
-  status?: 'idle' | 'loading' | 'success' | 'error';
-  setStatus?: ( status: 'idle' | 'loading' | 'success' | 'error' ) => void;
-}
-
-
-export interface WaitlistResponse_Beta {
-  waitlist: WaitlistEntryData;
-  createWaitlist: ( { email, name, newsletter, role }: WaitlistEntryData ) => Promise<boolean>;
-}
